pyproject.toml
[tool.poetry]
name = "enkaiplanner"
version = "0.1.0"
description = ""
authors = ["hiroki-yoda <hiroki-yoda@sensyn-robotics.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
cvxpy = "^1.4.2"
pandas = "^2.0.3"
openpyxl = "^3.0.2"
hydra-core = "^1.3.2"
flask = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
